+++ duperq/src/processor.rs
@@ -4,26 +4,27 @@ use smol::{
     channel,
     io::{AsyncWrite, AsyncWriteExt},
 };
+use yoke::Yoke;
 
 use crate::filter::DuperFilter;
 
 #[async_trait(?Send)]
 /// An opaque layer that processes a [`DuperValue`] asynchronously.
 pub trait Processor {
-    async fn process(&mut self, value: DuperValue<'static>);
+    async fn process(&mut self, value: Yoke<DuperValue<'static>, String>);
 
     async fn close(&mut self) {}
 }
 
 pub(crate) struct FilterProcessor {
     filter: Box<dyn DuperFilter>,
-    sender: channel::Sender<DuperValue<'static>>,
+    sender: channel::Sender<Yoke<DuperValue<'static>, String>>,
     is_open: bool,
 }
 
 impl FilterProcessor {
     pub(crate) fn new(
-        sender: channel::Sender<DuperValue<'static>>,
+        sender: channel::Sender<Yoke<DuperValue<'static>, String>>,
         filter: Box<dyn DuperFilter>,
     ) -> Self {
         Self {
@@ -36,8 +37,9 @@ impl FilterProcessor {
 
 #[async_trait(?Send)]
 impl Processor for FilterProcessor {
-    async fn process(&mut self, value: DuperValue<'static>) {
-        if self.is_open && self.filter.filter(&value) && self.sender.send(value).await.is_err() {
+    async fn process(&mut self, value: Yoke<DuperValue<'static>, String>) {
+        if self.is_open && self.filter.filter(value.get()) && self.sender.send(value).await.is_err()
+        {
             self.is_open = false;
         }
     }
@@ -45,18 +47,21 @@ impl Processor for FilterProcessor {
 
 pub(crate) struct TakeProcessor {
     available: usize,
-    sender: channel::Sender<DuperValue<'static>>,
+    sender: channel::Sender<Yoke<DuperValue<'static>, String>>,
 }
 
 impl TakeProcessor {
-    pub(crate) fn new(sender: channel::Sender<DuperValue<'static>>, available: usize) -> Self {
+    pub(crate) fn new(
+        sender: channel::Sender<Yoke<DuperValue<'static>, String>>,
+        available: usize,
+    ) -> Self {
         Self { sender, available }
     }
 }
 
 #[async_trait(?Send)]
 impl Processor for TakeProcessor {
-    async fn process(&mut self, value: DuperValue<'static>) {
+    async fn process(&mut self, value: Yoke<DuperValue<'static>, String>) {
         if self.available > 0 {
             if self.sender.send(value).await.is_err() {
                 self.available = 0;
@@ -72,12 +77,15 @@ impl Processor for TakeProcessor {
 
 pub(crate) struct SkipProcessor {
     to_skip: usize,
-    sender: channel::Sender<DuperValue<'static>>,
+    sender: channel::Sender<Yoke<DuperValue<'static>, String>>,
     is_open: bool,
 }
 
 impl SkipProcessor {
-    pub(crate) fn new(sender: channel::Sender<DuperValue<'static>>, to_skip: usize) -> Self {
+    pub(crate) fn new(
+        sender: channel::Sender<Yoke<DuperValue<'static>, String>>,
+        to_skip: usize,
+    ) -> Self {
         Self {
             sender,
             to_skip,
@@ -88,7 +96,7 @@ impl SkipProcessor {
 
 #[async_trait(?Send)]
 impl Processor for SkipProcessor {
-    async fn process(&mut self, value: DuperValue<'static>) {
+    async fn process(&mut self, value: Yoke<DuperValue<'static>, String>) {
         if self.to_skip > 0 {
             self.to_skip = self.to_skip.saturating_sub(1);
         } else if self.is_open && self.sender.send(value).await.is_err() {
@@ -99,11 +107,14 @@ impl Processor for SkipProcessor {
 
 pub(crate) struct OutputProcessor<O> {
     output: O,
-    printer: Box<dyn FnMut(DuperValue<'static>) -> Vec<u8>>,
+    printer: Box<dyn FnMut(Yoke<DuperValue<'static>, String>) -> Vec<u8>>,
 }
 
 impl<O> OutputProcessor<O> {
-    pub(crate) fn new(output: O, printer: Box<dyn FnMut(DuperValue<'static>) -> Vec<u8>>) -> Self {
+    pub(crate) fn new(
+        output: O,
+        printer: Box<dyn FnMut(Yoke<DuperValue<'static>, String>) -> Vec<u8>>,
+    ) -> Self {
         Self { output, printer }
     }
 }
@@ -113,7 +124,7 @@ impl<O> Processor for OutputProcessor<O>
 where
     O: AsyncWrite + Unpin + 'static,
 {
-    async fn process(&mut self, value: DuperValue<'static>) {
+    async fn process(&mut self, value: Yoke<DuperValue<'static>, String>) {
         self.output
             .write_all((self.printer)(value).as_ref())
             .await