+++ duperq/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Use `yoke` to prevent unnecessary cloning of Duper values.
+
 ## 0.1.2 (2025-12-23)
 
 ### Changed